requirements.txt
matplotlib==3.6.2
numpy==1.23.5
PyYAML==6.0
torch==1.13.1
tqdm==4.64.1
tensorboard==2.11.0
